tilføjede API mappe med et endpoint som laver GET af alle noter
